add support for typescript
