+++ package.json
@@ -12,6 +12,7 @@
     "@babel/core": "^7.4.4",
     "@babel/preset-env": "^7.4.4",
     "@types/node": "^11.13.8",
+    "@types/webpack": "^4.32.1",
     "babel-loader": "^8.0.5",
     "css-loader": "^2.1.1",
     "file-loader": "^3.0.1",
@@ -29,6 +30,7 @@
     "resolve-url-loader": "^3.1.0",
     "sass-loader": "^7.1.0",
     "style-loader": "^0.23.1",
+    "ts-loader": "^6.0.4",
     "ts-node": "^8.1.0",
     "typescript": "^3.4.5",
     "webpack": "^4.30.0",
