+++ tests/3/tsconfig.json
@@ -1,0 +1,3 @@
+{
+	"extends": "../../typescript/tsconfig.strict.json"
+}
