+++ typescript/tsconfig.base.json
@@ -1,0 +1,6 @@
+{
+	"compilerOptions": {
+		"target": "es3",
+		"sourceMap": true
+	}
+}
