+++ typescript/tsconfig.strict.json
@@ -1,0 +1,7 @@
+{
+	"extends": "./tsconfig.base.json",
+	"compilerOptions": {
+		"noImplicitAny": true,
+		"strictNullChecks": true
+	}
+}
